requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
PyPDF2
pinecone-client
numpy
scikit-learn
langchain
together
langchain_together
langchain_community
fpdf